+++ app/code/pagetype/Page.php
@@ -1,7 +1,41 @@
 <?php
 
 use SilverStripe\CMS\Model\SiteTree;
+use SilverStripe\Forms\TextareaField;
 
 class Page extends SiteTree
 {
+    /**
+     * Page database name
+     *
+     * @var string
+     */
+    private static $table_name = "PageField";
+
+    /**
+     * Page database
+     *
+     * @var array
+     */
+    private static $db = [
+        'Summary' => 'Text'
+    ];
+
+    /**
+     * Create fields in the page settings of the CMS
+     *
+     * @return void
+     */
+    public function getCMSFields()
+    {
+        $fields = parent::getCMSFields();
+        $fields->addFieldToTab(
+            'Root.Main',
+            TextareaField::create(
+                'Summary'
+            ),
+            'Content'
+        );
+        return $fields;
+    }
 }
