+++ themes/app/src/scss/_base.scss
@@ -18,14 +18,3 @@ p {
     margin-left: 0%;
     margin-right: 0%;
 }
-
-.page-content__main {
-    img {
-        max-width: 100%;
-        height: auto;
-    }
-
-    p {
-        margin: 20px 0;
-    }
-}
