+++ themes/app/src/scss/_typography.scss
@@ -8,3 +8,14 @@
 body {
     font-family: Federant, serif;
 }
+
+.typography {
+    img {
+        max-width: 100%;
+        height: auto;
+    }
+
+    p {
+        margin: 20px 0;
+    }
+}
