--- themes/app/templates/App/PageType/Layout/ContentPage.ss
@@ -1,12 +1,0 @@
-<% include Breadcrumb %>
-<div class="container">
-    <div class="row">
-        <div class="page-content__main col-lg-8">
-            <p class="lead page-content__summary">$Summary</p>
-            $Content
-        </div>
-        <div class="offset-xl-1 col-xl-3 col-lg-4">
-            <%-- <% include Alert %> --%>
-        </div>
-    </div>
-</div>
