--- themes/app/templates/Includes/Breadcrumb.ss
@@ -1,5 +1,0 @@
-<div class="breadcrumb">
-    <div class="container">
-        $Breadcrumbs
-    </div>
-</div>
